ASYLUM: enhance "objects" console command
this command now takes a format option of "idx" or "id" so you can look up an object entry by its internal id as well as it's physical location within the objects array.

Also enhanced Objects::toString with an optional formatting parameter that displays a short dump or a long dump.

sample usage:
  
  objects idx 51   // summary view of object at array index 51
  objects id 1258  // detail view of object with id of 1258
  objects          // show all objects (summary list)

git-svn-id: https://asylumengine.googlecode.com/svn/trunk@496 0bfb4aae-4ea4-11de-8d8d-752d95cf3e3c